Implement the client login into the server
Also, we use a RwLock to ensure that only one tokio thread can modify
the state variable.

A RwLock<> was chosen because it was the only one that I could write
without having to create a mutable object, and if I created a mutable
mutex or Arc, I would not be able to use the variable into a spawned
thread.