Add a set moderator endpoint that will only be called once
Every client start as a normal client, so we need a way to set a
client to moderator.

I chose it by a password. In the future, this server will be launched
from the main executable and the password will be passed as a
parameter, or as an environment variable, and the host client will be
the first to log in and guess the password.

For now, this method seems the most cleaner, and the better given the
conditions (of the client having to depend on something from the
server)